Libft V6 final

+++ Libft/ft_strchr.c
@@ -6,15 +6,15 @@
 /*   By: eguelin <eguelin@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 16:04:38 by eguelin           #+#    #+#             */
-/*   Updated: 2022/11/20 13:52:32 by eguelin          ###   ########lyon.fr   */
+/*   Updated: 2022/11/25 13:13:21 by eguelin          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
+#include "libft.h"
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	pos;
+	size_t	pos;
 
 	pos = 0;
 	while (s[pos])
